Fix some problems with rangearmor
Some problems were still reported... one of them is that the copying of some files is not successful if done within the range engine.
it turns out that when trying to use the range engine it took the path of the engine and not the executable (rangearmor.exe) causing a error.
now it is able to operate in both situations.
- Example Game Updated
- RangeArmor now uses GLES3

Tested on Windows 64 and Linux 64.