RangeArmor: Major changes, deprecate obfuscation system
The main purpose of the tool (formerly BGArmor) was to organize and obfuscate the games released on the gameengine. But the obfuscator is doing more harm than good; Offers security flaws after starting the game, problems with persistent files if the game is closed incorrectly (like a power outage, for example) does not save files which causes the player to lose progress if it has a save system or other things, increases the initial loading time of games, and sometimes fails to flush temporary files.
Currently, the range engine protected file provides good protection for games, so now the files are directly stored in the data folder.

- Deprecate linux x32, now rangearmor no longer supports linux x32.
- Godot updated to 3.5.2.
- RangeArmor bump to 1.4.0.
- UI improvements.
- Fix does not export launcher for windows.

Tested on windows only, partial linux support added (ToDo).